backend/requirements.txt
# FastAPI
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.2.1
google-generativeai==0.8.6
